Sega. SOD test with second order reconstruction (JB limiter) is runnable. All parameters and method are the same as in the old code:
https://github.com/sergantche/NewKidzTurboOld/commit/a4c811acfbcf5335b0edeb71fd6dd8ce1cd21cf4

+++ src/utility/Limiters.h
@@ -17,7 +17,7 @@ public:
 	std::valarray<double> lim_values;
 
 	// get limiters value vor i-th variable
-	int inline operator[] (int i) {
+	double inline operator[] (int i) {
 		return lim_values[i];
 	};
